--bug=90236325 管理员无法查看私有项目的流水线执行数据

+++ src/backend/ci/ext/tencent/project/api-project-op/src/main/kotlin/com/tencent/devops/project/api/op/OpUserResource.kt
@@ -78,6 +78,6 @@ interface OpUserResource {
         limit: Int?,
         @ApiParam("间隔时间", required = false)
         @QueryParam("sleep")
-        sleep: Int?
+        sleep: Long?
     ): Result<Int>
 }
